Statement-2-Lucknow-Foodie/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
chromadb==0.4.24
pandas==2.2.1
pydantic==2.6.4
python-dotenv==1.0.1
openai==1.14.0
